lxqtloosefocusframe and lxqttaskgroup splited into seperate files
