+++ plugin-taskbar/lxqttaskgroup.cpp
@@ -12,7 +12,7 @@
 #include <QDragLeaveEvent>
 #include <QMenu>
 #include <XdgIcon>
-#include <QPropertyAnimation>
+#include "lxqttaskpopup.h"
 
 LxQtTaskGroup::LxQtTaskGroup(const QString &groupName,QIcon icon,ILxQtPanelPlugin * plugin, LxQtTaskBar *parent):
     LxQtTaskButton(0,parent,parent),
@@ -63,9 +63,7 @@ void LxQtTaskGroup::contextMenuEvent(QContextMenuEvent *event)
     }
 
     QMenu menu(tr("Group"));
-
     menu.addAction(XdgIcon::fromTheme("process-stop"), tr("Close group"),this,SLOT(closeGroup()));
-
     menu.exec(mapToGlobal(event->pos()));
 }
 
@@ -252,9 +250,11 @@ void LxQtTaskGroup::removeButton(WId window)
         }
         else
         {
+            if (isVisible())
+                emit visibilityChanged(false);
             hide();
             groupBecomeEmpty(groupName());
-            emit visibilityChanged(false);
+
         }
     }
 }
@@ -629,182 +629,11 @@ void LxQtTaskGroup::windowChanged(WId window, NET::Properties prop, NET::Propert
 
     if (prop.testFlag(NET::WMState))
         button->setUrgencyHint(KWindowInfo(window, NET::WMState).hasState(NET::DemandsAttention));
-
 }
 
 /************************************************
 
  ************************************************/
-
-
-/************************************************
-
- ************************************************/
-LxQtLooseFocusFrame::LxQtLooseFocusFrame(const QHash<WId,LxQtTaskButton* > & buttons,QWidget *parent):
-    QDialog(parent),
-    mButtonHash(buttons)
-{
-    //setWindowFlags( Qt::Dialog | Qt::WindowStaysOnTopHint | Qt::CustomizeWindowHint | Qt::Popup | Qt::X11BypassWindowManagerHint);
-    setWindowFlags(  Qt::CustomizeWindowHint | Qt::FramelessWindowHint | Qt::ToolTip);
-//    mFrame->setAttribute(Qt::WA_X11NetWmWindowTypeDock);
-    setAttribute(Qt::WA_AlwaysShowToolTips);
-    //setMouseTracking(true);
-    setAcceptDrops(true);
-
-    mPosAnimation = new QPropertyAnimation(this,"pos",this);
-    mPosAnimation->setDuration(200);
-
-    mSizeAnimation = new QPropertyAnimation(this,"size",this);
-    mSizeAnimation->setDuration(200);
-}
-/************************************************
-
- ************************************************/
-LxQtLooseFocusFrame::~LxQtLooseFocusFrame()
-{
-}
-
-/************************************************
-
- ************************************************/
-void LxQtLooseFocusFrame::enterEvent(QEvent *event)
-{
-    QEvent::Type t = event->type();
-    if (t == QEvent::Enter)
-    {
-        mouseLeft(false);
-    }
-
-    QDialog::enterEvent(event);
-}
-
-/************************************************
-
- ************************************************/
-void LxQtLooseFocusFrame::leaveEvent(QEvent *event)
-{
-    QEvent::Type t = event->type();
-    if(t == QEvent::Leave)
-    {
-        emit mouseLeft(true);
-    }
-
-    QDialog::leaveEvent(event);
-}
-
-/************************************************
-
- ************************************************/
-void LxQtLooseFocusFrame::dropEvent(QDropEvent *event)
-{
-    buttonDropped(event->pos(),event);
-}
-
-/************************************************
-
- ************************************************/
-void LxQtLooseFocusFrame::dragEnterEvent(QDragEnterEvent *event)
-{
-    if (!event->mimeData()->hasFormat("lxqt/lxqttaskbutton"))
-    {
-        event->ignore();
-        return;
-    }
-    if (mButtonHash.count() == 1)
-    {
-        event->ignore();
-        return;
-    }
-
-    event->acceptProposedAction();
-
-    //LxQtTaskButton::dragEnterEvent(event);
-    QWidget::dragEnterEvent(event);
-}
-
-/************************************************
-
- ************************************************/
-void LxQtLooseFocusFrame::buttonDropped(const  QPoint& point, QDropEvent *event)
-{
-    WId window;
-    QDataStream stream(event->mimeData()->data("lxqt/lxqttaskbutton"));
-    stream >> window;
-    if (!mButtonHash.contains(window))
-    {
-        return;
-    }
-
-    LxQtTaskButton * dragged = mButtonHash.value(window);
-    int droppedIndex = layout()->indexOf(dragged);
-    int newIdx = -1;
-    int temp;
-
-    int oldTreshold = 0;
-    for (int i = 0 ; i < layout()->count(); i++)
-    {
-        QWidget * w = layout()->itemAt(i)->widget();
-        LxQtTaskButton * b = qobject_cast<LxQtTaskButton*>(w);
-        if (b && w->isVisibleTo(this))
-        {
-            int treshold = b->pos().y() + b->height() ;
-            if (oldTreshold <= point.y() && point.y() < treshold)
-            {
-                newIdx = i;
-                break;
-            }
-            temp = i;
-            oldTreshold = treshold;
-        }
-    }
-
-    /*
-    if (newIdx == -1)
-        newIdx = temp+ 1;
-        */
-
-    QVBoxLayout * l = qobject_cast<QVBoxLayout *>(layout());
-    l->insertWidget(newIdx,dragged);
-}
-
-void LxQtLooseFocusFrame::moveEyeCandy(const QPoint  & newPos)
-{
-    LxQtTaskGroup* group = qobject_cast<LxQtTaskGroup*>(parent());
-    Q_ASSERT(group);
-    bool eyecandy = group->parentTaskBar()->settings().eyeCandy;
-
-    if (eyecandy)
-    {
-        mPosAnimation->stop();
-        mPosAnimation->setStartValue(pos());
-        mPosAnimation->setEndValue(newPos);
-        mPosAnimation->start();
-    }
-    else
-    {
-        move(newPos);
-    }
-}
-
-void LxQtLooseFocusFrame::resizeEyeCandy(int w, int h)
-{
-    LxQtTaskGroup* group = qobject_cast<LxQtTaskGroup*>(parent());
-    Q_ASSERT(group);
-    bool eyecandy = group->parentTaskBar()->settings().eyeCandy;
-
-    if (eyecandy)
-    {
-        mSizeAnimation->stop();
-        mSizeAnimation->setStartValue(size());
-        mSizeAnimation->setEndValue(QSize(w,h));
-        mSizeAnimation->start();
-    }
-    else
-    {
-        resize(w,h);
-    }
-}
-
